requirements.txt
astroid==2.4.2
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
decorator==4.4.2
idna==2.10
isort==4.3.21
joblib==0.16.0
JPype1==0.7.5
konlpy==0.5.2
lazy-object-proxy==1.4.3
lxml==4.5.1
mccabe==0.6.1
networkx==2.4
numpy==1.19.0
oauthlib==3.1.0
pandas==1.0.5
pylint==2.5.3
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
scikit-learn==0.23.1
scipy==1.5.0
six==1.15.0
soupsieve==2.0.1
threadpoolctl==2.1.0
toml==0.10.1
tweepy==3.8.0
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.12.1
